+++ client/web/src/insights/pages/dashboards/dashboard-page/components/dashboards-content/hooks/use-dashboard-select-handler.ts
@@ -3,11 +3,13 @@ import { useHistory } from 'react-router-dom'
 import { InsightDashboard, isVirtualDashboard } from '../../../../../../core/types'
 import { isSettingsBasedInsightsDashboard } from '../../../../../../core/types/dashboard/real-dashboard'
 
+type SelectHandler = (dashboard: InsightDashboard) => void
+
 /**
  * Hook for managing URL of the dashboard page whenever the user picks
  * another dashboard via dashboard select
  */
-export function useDashboardSelectHandler() {
+export function useDashboardSelectHandler(): SelectHandler {
     const history = useHistory()
 
     return (dashboard: InsightDashboard): void => {